requirements.txt
pandas
python-dotenv
black
pytest
sqlalchemy
psycopg2-binary